+++ src/io/load_case.jl
@@ -37,7 +37,7 @@ function load_case(input_dir::String, work_dir::String, case_name::String, base_
 
     # parse data
     sn_data = parse_file(m_file)  # Parse initial single-network structure
-    hours = length(find_value(time_series))  # Find number of simulated hours (a single dimension is used)
+    hours = find_hours(time_series)
     _FP.add_dimension!(sn_data, :hour, hours)  # Add hour dimension to single-network data
     mn_data = _FP.make_multinetwork(sn_data, time_series; share_data=false)  # Create the multinetwork data dictionary
     _PMMCDC.make_multiconductor!(mn_data)  # Convert the DC grid (if any) of each single-network model to multi-conductor
@@ -103,4 +103,22 @@ function build_time_series(work_dir::String, base_mva::Int=100)
     end
 
    return time_series
+end
+
+function find_hours(time_series)
+    n_hours = 0
+    error_message_data = ""
+    for comp_name in keys(time_series)
+        for comp_id in keys(time_series[comp_name])
+            for attribute in keys(time_series[comp_name][comp_id])
+                hours = length(time_series[comp_name][comp_id][attribute])
+                if n_hours != 0
+                    @assert hours == n_hours "time_series[$comp_name][$comp_id][$attribute] has data for $hours hours, whil $error_message_data has data for $n_hours"
+                end
+                n_hours = hours
+                error_message_data = "time_series[$comp_name][$comp_id][$attribute]"
+            end
+        end
+    end
+    return n_hours
 end