+++ src/io/run_study.jl
@@ -1,14 +1,24 @@
 using Distributed
 using HVDCWISE_TEA
-import HVDCWISE_TEA as _HWTEA
+import PowerModels as _PM
 import Ipopt
 
-const _HWTEA_dir = dirname(dirname(pathof(HVDCWISE_TEA))) # Root directory of HVDCWISE_TEA package
+const _HWTEA_dir = dirname(dirname(pathof(HVDCWISE_TEA)));  # Root directory of HVDCWISE_TEA package
+
+## Workers setup
+
+nprocs() > 1 && rmprocs(workers())
+addprocs(Sys.CPU_THREADS ÷ 2; exeflags = "--project=$(Base.active_project())")
+@everywhere begin
+    using HVDCWISE_TEA
+    using Ipopt
+    HVDCWISE_TEA.silence()
+end
 
 
-_PM.silence()
+# import HVDCWISE_TEA as _HWTEA
 
-function run_study(work_dir::String, case_name::String)
+function run_study_old(work_dir::String, case_name::String)
     base_mva = 100
     # build_raw_inputs(work_dir) -> the .m at the root and 1 folder per micro with the .csv
     build_raw_inputs(work_dir, work_dir, case_name, base_mva, true)
@@ -29,20 +39,46 @@ function run_study(work_dir::String, case_name::String)
     #
 end
 
-# work_dir = joinpath(_HWTEA_dir, "studies\\IEEE39\\3_hours_with_storage")
-# case_name = "IEEE39"
-# @time run_study(work_dir, case_name)
+# @time run_study(joinpath(_HWTEA_dir, "studies\\IEEE39\\3_hours_without_storage"), "IEEE39")
 
-# work_dir = joinpath(_HWTEA_dir, "studies\\IEEE39\\draft")
-# case_name = "test05b"
-# @time load_case(work_dir, work_dir, case_name, 100, true)
+function run_study(work_dir::String, hours::Int, base_mva::Int=100)
+    println("Build simulation inputs")
+    build_simulation_inputs(work_dir, base_mva)
 
-@time run_study(joinpath(_HWTEA_dir, "studies\\IEEE39\\3_hours_without_storage"), "IEEE39")
+    println("Run simulation")
+    #Paths
+    simulation_dir = joinpath(work_dir, "simulation_interface")
+    path2grid = find_grid_path(simulation_dir)
+    path2data = joinpath(simulation_dir, "Input_series")
 
-# @time run_study_2("draft", "IEEE39")
-# @time run_study("3_hours_without_storage", "IEEE39")
-# run_study("1_day_without_storage", "IEEE39")
-# run_study("1_day_with_storage", "IEEE39")
-# @time run_study("1_day_with_storage", "IEEE39")
-# @time run_study("1_week_with_storage", "IEEE39")
-# @time run_study("1_year_with_storage", "IEEE39")
+    ## Solver parameters
+    optimizer = HVDCWISE_TEA.optimizer_with_attributes(Ipopt.Optimizer, "print_level" => 0)
+    setting = Dict("output" => Dict("branch_flows" => true, "duals" =>false), "conv_losses_mp" => false);
+    
+    ## Solve the multiperiod OPF problem
+    run_tea(path2grid, path2data, hours, _PM.DCPPowerModel, optimizer; setting = setting)
+
+    println("Simulation finished")
+    # TODO add post-processing here
+end
+
+function find_grid_path(simulation_dir::String)
+    matpower_name = ""
+    for file_name in readdir(simulation_dir)
+        file_path = joinpath(simulation_dir, file_name)
+        if isfile(file_path)
+            macro_scenario = file_name[1:length(file_name)-2]
+            if file_name == "$macro_scenario.m"
+                @assert (matpower_name == "")  "2 files could contain the grid model data in $simulation_dir: $macro_scenario.m and $matpower_name. Please delete or rename the wrong file."
+                matpower_name = file_name
+            end
+        end
+    end
+    @assert (matpower_name != "")  "No matpower file has been found in $simulation_dir."
+    return joinpath(simulation_dir, matpower_name)
+end
+
+work_dir = joinpath(_HWTEA_dir, "studies\\simple_use_case")
+hours_per_subsimulation = 96  # The yearly problem is split into subproblems of this size (1 week if 168h)
+
+run_study(work_dir, hours_per_subsimulation)